Add `flightSearch` to FlightAPI
Update the Flight API prototype to include a new method that accepts an
airline code and a set of parameters to query the external Locomotive
code task server for results. This is independent of the final
collection of search results for all given airline codes and only works
for one flight code, but should used multiple times to bring those
results together.

== Notes:
- Add appropriate test, whilst mocking out external HTTP calls

- Add new `querystring` library to encode JSON object to parameters for
  the final query URL

== References:
- [Query String | Node.js v8.0.0 Documentation]
  (https://nodejs.org/api/querystring.html)

- [javascript set a variable if undefined - Stack Overflow]
  (https://stackoverflow.com/questions/5409641/javascript-set-a-variable-if-undefined)`