Update SearchForm to hold properties NOT state
Update the Search Form Component to hold properties and NOT state, since
it will be later used within a parent container component to render
their state as this SearchForm component properties, as well as a list
of results. Also, since it is not changing state it will instead just be
re-rendered via the parent component state is the wrong word to describe
this behaviour and the code should reflect that (as well as in tests).

Add the appropriate input values as properties that can be assigned to
the component, render them within Document Object Model (DOM) and update
the tests accordingly to render empty values on initial render or
subsequent re-renders.

== Notes:
- Since the values will also be empty on initial render this would be a
  good check to run validation or not, unless a flag some sort; e.g.
  `reRendered`, `runValidation` or `validate`; gets set between renders
  most likely caused by the `findSearchResults` callback function being
  used to render search results (or if no search results found then run
  validation?)