Add WIP SearchBox component
Add the Work In Progress (WIP) Search Box component that builds a
container component with internal state to represent the final rendered
component within the application `index.html` view`. The reason its
still a WIP is that are missing tests or incomplete tests (despite most
of the mocking of remote API calls being complete) however unexpectedly
the library `nock` used for mocking external API calls is breaking other
tests (_really?_), which is a shame and requires that mocks used to be
similar or the same if the same URL being mocked is used; e.g.
`/airlines`.

== Notes:
- Add some, but not all, appropriate tests

- Update Flight API specification tests to fix breaking same mocked
  `nock` requests; `/airports` and `/airlines`

== References:
- [react/ReactBaseClasses.js at 5c6a496d98b80d19d851c2ec3a56d760b7118a50
  · facebook/react]
  (https://github.com/facebook/react/blob/5c6a496d98b80d19d851c2ec3a56d760b7118a50/src/isomorphic/modern/class/ReactBaseClasses.js#L59)

- [react/ReactMount.js at b1b4a2fb252f26fe10d29ba60d85ff89a85ff3ec ·
  facebook/react]
  (https://github.com/facebook/react/blob/b1b4a2fb252f26fe10d29ba60d85ff89a85ff3ec/src/renderers/dom/stack/client/ReactMount.js)

- [The Jasmine Spy Cheatsheet]
  (http://tobyho.com/2011/12/15/jasmine-spy-cheatsheet/)