Fix `_searchFromServer` settings during tests
Fix the relative path and domain settings used within the SearchBox
component in tests (and to a lesser extent in the browser for cross-site
scripting purposes) in order to provide some defaults for the domain URL
and search Path used to build the necessary Get request from the
localhost server to preform the flight search. Otherwise the Node
XMLHTTPRequest library will not default to localhost during tests and
fail to resolve the request successful; thus failing the test.