Fix XMLHTTPRequest URL within SearchBox component
Fix the parameters used within the Search Box component in order to
properly query the localhost server running the back-end code, otherwise
the results will always be empty and fail the test. This resolves that
missing data problem and also only updates the component on successful
response, otherwise it should return an error code or message in the
future due to a bad request or no results (not yet implemented).

== Notes:
- Update `_getResults (after AJAX request with params)` test to
  piggyback off the injected callback `reRenderCallback` in order to
  inspect the Search Box component state before / as it gets re-rendered
  since by that time the XMLHTTPRequest will have been finished and the
  callback will be called

- Update the location of the `sydneyToMelbourneFlightSearchData` so that
  it can be compare expected mock data with the actual mock, despite
  being the same data, the around about way it is reached; via the
  back-end code querying the remote API that has been stubbed out; is
  what is actually being tested

- Add a comment about the callback function within `_searchFromServer`
  function, to use instead of piggy backing of the re-render call and to
  make it possible for something else to be done with the returned data
  outside the updating of the Component