Fix `search` mistake in wrong Airports assumption
Fix a problem with all the tests that use the `search` function within
the FlightAPI service, since it mistakenly assumed that the search would
receive the airport codes directly and not actually the location of the
airports; e.g. assumed `'MLB'` instead the actual expected input
`'Melbourne'` for the `from` and `to` fields within searches.

== Notes:
- Add some light documentation to better describe the expected
  parameters for the `flightSearch` and `search` to help clarify this
  mistake

== References:
- [Use JSDoc: @param]
  (http://usejsdoc.org/tags-param.html)