Add FlightAPI `airportCodesFromLocation` function
Add the Airport Codes from Location function to the Flight API in order
to provide a function that can return a collection of required airport
codes for a given location. It will be later used to convert the given
location name to an array of Airport Code values; e.g.

    'Melbourne' -> ['MLB', 'MEL']
    'Sydney'    -> ['SYD', 'YQY']

That will later be used within the final FlightAPI `search` function to
query the remote Locomotive code task API.

== Notes:
- Add some appropriate tests

- Add some light documentation to help differentiate the different
  functions

- Update Notes to admit broken assumption the logic to help fix