Add missing mocked flight search URLs for tests
Update the FlightAPI specification test setup for the `search` function
test to include previously missing airline code and travel parameters
URLs to match all the expected flight search URL requests being made in
a single search. Also, set them respond with `404` Not Found errors in
order reduce the amount of data being return in the final result for the
purposes of these tests.