Update & Fix `search` function in FlightAPI
Update and Fix the `search` function within the Flight API service to
remove the previously built-in assumption that Users of the search
function would provide the Airport codenames as the `from` & `to`
variables instead of their locations; e.g. 'MEL' vs 'Melbourne' or 'YGY'
vs 'Sydney'. This resolves this mistake by first querying a given
location using `airports?q=` for both the `from` and `to` locations
before continuing to build the final flight search query using
`asyncFlightSearch` function, for every `airline` and `airport`
combinations.

== Notes:
- The FlightAPI tests should now be passing

- Fixes another implicit mistake or concern made with using the `params`
  variable within the old inlined `airlineFlightSearch` function  that
  was created due to the original mistake of not querying a location for
  airport codes