Add curried Flight Search as an alternative search
Add an Asynchronous curried Flight Search alternative to build up the
final Flight Search request within the Search function, which was
initially built to help debug the error of where the bug within the
non-async `flightSearch` not handling `404` errors, but was also useful
to compare to the final solution in terms of complexity.

Whilst this curried Flight Search works it has some flaws, such as
needing to pass the Flight Search data each time it resolves the
Promise, due to Promises only being able to return a single value and
then retrieving it for each subsequent call so that the final
`travelDate` data, as well as other various collections, are available
at the end of `search` function.

The `curriedFlightSearch` function simply returns a Promise containing a
function with the next link in the chain of functions returning
functions to build the final `flightSearch` query, passing each
parameter one at a time, returning an incomplete function for each value
in the arrays of information require of the final search calls.

== References:
- [Array.prototype.reduce() - JavaScript | MDN]
  (https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/Reduce?v=control)

- [Currying in JavaScript – Kristina Brainwave – Medium]
  (https://medium.com/@kbrainwave/currying-in-javascript-ce6da2d324fe)

- [A Beginner's Guide to Currying in Functional JavaScript — SitePoint]
  (https://www.sitepoint.com/currying-in-functional-javascript/)

- [Currying in JavaScript ES6 – Adam Bene – Medium]
(https://medium.com/@adambene/currying-in-javascript-es6-540d2ad09400)