Update some `nock` requests to use `query`
Update some of the Nock `get` requests to make use of the `query` chain
method in order to provide an alternative way of providing the same URL
query. However, only apply it to some to ensure that the present URLs
still work and will still work between tests in the future.

== References:
- [node-nock/nock: HTTP mocking and expectations library]
  (https://github.com/node-nock/nock/#specifying-request-query-string)