Add 400 error mock data for SearchBox component
Add additional 400 return error code as mock data for several requests
made within the callback in the Search Box component, since the fix
applied to the server back-end that expects the Airport locations for
the `from` & `to` inputs, not their actual Airport codes that was
mistaken assumed. This means it is inspecting all available airports for
a given location, hence more requests that need to be stubbed out.