Add RequiredValidator to validate a value presence
Add a Required Validator class to hold a validate function that accepts
a context, a property name and that properties value as parameters. It
then inspects the value to ensure that it is not null, undefined or an
empty string, otherwise it adds an error message to the context's error
collection object under the name of the property to explain to an user
that the value is required.

== Notes:
- A parent object; i.e. context; can then later determine if it is
  indeed invalid by the number of errors it has; e.g. 0 error keys means
  its valid, otherwise more than 0 means it is invalid

- This validator is to be used with Input Components later on to
  determine if a value is valid when submitted via a form (or within the
  backend if need be)

== References:
- [ActiveModel::Validator]
  (http://api.rubyonrails.org/classes/ActiveModel/Validator.html)

- [Active Record Validations — Ruby on Rails Guides]
  (http://guides.rubyonrails.org/active_record_validations.html#custom-validators)

- [null \- How do you check for an empty string in JavaScript? \- Stack Overflow]
  (https://stackoverflow.com/questions/154059/how-do-you-check-for-an-empty-string-in-javascript)