Add disabled option to SearchForm to disable input
Add a disabled boolean option to the Search Form Component in order to
disabled inputs (and form submission) when it is set to be disabled.
This will be later used to prevent the User from interacting with the
Form during its submission and spawn multiple HTTP requests, whilst it
is loading the previous results in the meantime.