Add initial draft of Paginator Component
Add an initial rough draft of the Paginator and Pagination Components to
encapsulate the idea of pages of content, with a collection of links to
click between the different pages.

However, the on click event for these blank anchor links has yet been
handled appropriately and may required a call to re-render the parent
Component, whilst updating the current index to the value of the link
index value.

== References:
- [How To Make a Pagination]
  (https://www.w3schools.com/howto/howto_css_pagination.asp)

- [kaminari/_paginator.html.erb at master · kaminari/kaminari]
  (https://github.com/kaminari/kaminari/blob/master/kaminari-core/app/views/kaminari/_paginator.html.erb)