Update Search Result layout & visual CSS
Update the Search Result layout to better space out vertically stacked
elements and spread out the information of the journey between airports,
as well as the distance / time between them. Also, emphasis the price /
fare of the flight from the rest of the result, since it is often the
most important piece of information that Customer cares about.

== Notes:
- Fix inconsistencies with colours used to highlight buttons (from green
  to blue)